+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 - Fix(tui): properly set lyric content on reload (like exiting the config editor).
 - Fix(tui): populate "Stopped" state to everything instead of showing stale data.
 - Fix(tui): change how progress time is displayed to now be `--:--` if unavailable / stopped, and dont show total duration if unavailable.
+- Fix(tui): dont show error popup if `cover-ueberzug` is compiled-in but `ueberzug` command is not available.
 - Fix(server): on rusty backend, update symphonia to fix various issues.
 - Fix(server): on rusty backend, update `rusty-libopus` to dynamic link without custom environment variable.